Extract strongswan to a Driver in ipsec
In order to support other IPsec drivers, the Driver interface has been
introduced to encapsulate the requirements from the IPsec engine.
The IPsec engine has some business logic and iptables related code which
isn't related to the way it actually connects the tunnels, so that code
remains in the engine, while the tunnel specific things have been
extracted.

The implementation has been extracted as-is, with as little changes as
possible (other than the obligatory ones to actually allow the new
interface) so that the refactor is easy to understand.

Currently the StrongSwan implementation is created hard coded, but the
aim is to add a LibreSwan driver type in the next patches.
